refactor: dispatch keyb events and simplify func
Previously, the click event listeners and keyboard event listeners
handled everything separately, which was fine but makes it more
difficult to manage. Now the click event listeners do nothing but
dispatch keyboard events and the keyboard event listeners handle all
the logic.

Also the operate function was modified. Instead of returning one of the
four required functions, it now instead returns the calucalted values
which I felt is a little cleaner.