fix: handle operand logic for states 0, 2, 4 and 5
Previously, when in states 0, 2 or 5 operand1 instead of getting
intialized was getting appended. Same for operand 2 in state 4. This
meant for calculations where we are chaining operators, the previous
operands were still in the variables and new operands were just getting
appended to it.

Now this issue is fixed and in states 0, 2, 4 and 5, the operands are
no longer appended but rather intialized.